Refactor and enhance file serving plugins
- Remove old static-file-plugin-nodejs package
- Add comprehensive serve-static-plugin-nodejs with Range requests support
- Enhance send-file-plugin-nodejs with improved architecture and security
- Update changesets for both plugins
