+++ .changeset/serve-static-plugin-nodejs.md
@@ -1,0 +1,11 @@
+---
+'@korix/serve-static-plugin-nodejs': patch
+---
+
+Add serve-static-plugin-nodejs package for comprehensive static file serving
+
+- Static file serving with configurable mount points
+- Range requests support for streaming and partial content
+- Cache headers with ETag and Last-Modified support
+- Security features with path traversal protection
+- Index file resolution and dotfile handling
