--- .changeset/static-file-plugin-nodejs.md
@@ -1,13 +1,0 @@
----
-'@korix/static-file-plugin-nodejs': patch
----
-
-Add static file serving plugin for Node.js
-
-- Secure static file serving with path traversal protection
-- Streaming file delivery without buffering
-- MIME type detection and proper Content-Type headers
-- ETag and Last-Modified support for conditional requests
-- Range request support for partial content delivery
-- Cache-Control headers with optional immutable directive
-- Comprehensive README with usage examples and best practices
