+++ packages/send-file-plugin-nodejs/src/header/create-cache-control-header.ts
@@ -1,0 +1,13 @@
+export function createCacheControlHeader(maxAge?: number, immutable?: boolean): string | null {
+  const parts: string[] = [];
+
+  if (typeof maxAge === 'number') {
+    parts.push(`max-age=${maxAge}`);
+  }
+
+  if (immutable === true) {
+    parts.push('immutable');
+  }
+
+  return parts.length > 0 ? parts.join(', ') : null;
+}
