+++ packages/serve-static-plugin-nodejs/eslint.config.js
