+++ packages/serve-static-plugin-nodejs/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@korix/static-file-plugin-nodejs",
+  "name": "@korix/serve-static-plugin-nodejs",
   "version": "0.1.0",
-  "description": "Static file serving plugin for Kori framework (Node.js)",
+  "description": "Serve static files plugin for Kori framework (Node.js)",
   "type": "module",
   "author": "mitz (@bufferings)",
   "license": "MIT",
@@ -56,6 +56,7 @@
     "sync:version": "ks sync-version && prettier --write src/version.ts"
   },
   "dependencies": {
+    "etag": "1.8.1",
     "mime-types": "3.0.1"
   },
   "peerDependencies": {
@@ -65,6 +66,7 @@
     "@korix/eslint-config": "workspace:*",
     "@korix/kori": "workspace:*",
     "@korix/script": "workspace:*",
+    "@types/etag": "1.8.4",
     "@types/mime-types": "3.0.1",
     "@types/node": "catalog:",
     "eslint": "catalog:",
