+++ packages/serve-static-plugin-nodejs/src/index.ts
@@ -1,0 +1,2 @@
+export { type ServeStaticOptions, serveStaticPlugin } from './serve-static-plugin/index.js';
+export { PLUGIN_VERSION } from './version.js';
