+++ packages/serve-static-plugin-nodejs/src/serve-static-plugin/file/index.ts
@@ -1,0 +1,1 @@
+export { type FileInfo, resolveIndexFile } from './file-info.js';
