+++ packages/serve-static-plugin-nodejs/src/serve-static-plugin/index.ts
@@ -1,0 +1,2 @@
+export { type ServeStaticOptions } from './options.js';
+export { serveStaticPlugin } from './plugin.js';
