+++ packages/serve-static-plugin-nodejs/src/serve-static-plugin/options.ts
@@ -1,4 +1,4 @@
-export type StaticFileOptions = {
+export type ServeStaticOptions = {
   /** Required: Source directory path to serve static files from */
   serveFrom: string;
 
@@ -29,34 +29,3 @@ export type StaticFileOptions = {
   /** Maximum number of ranges per request (default: 1) */
   maxRanges?: number;
 };
-
-/**
- * Parsed range information
- */
-export type ParsedRange = {
-  start: number;
-  end: number;
-};
-
-/**
- * Range request result
- */
-export type RangeResult = {
-  ranges: ParsedRange[];
-  totalSize: number;
-  isSatisfiable: boolean;
-};
-
-/**
- * Range Request related constants
- */
-export const RangeConstants = {
-  /** Range unit for HTTP Range requests */
-  BYTES: 'bytes',
-
-  /** Accept-Ranges header value for no range support */
-  NONE: 'none',
-
-  /** Range header prefix */
-  RANGE_PREFIX: 'bytes=',
-} as const;
