+++ packages/serve-static-plugin-nodejs/src/serve-static-plugin/security/index.ts
@@ -1,0 +1,1 @@
+export { resolveSafePath } from './resolve-safe-path.js';
