+++ packages/serve-static-plugin-nodejs/src/serve-static-plugin/serve/index.ts
@@ -1,0 +1,2 @@
+export { handleStaticFileRequest } from './handle-request.js';
+export { serveFile } from './serve-file.js';
