+++ packages/serve-static-plugin-nodejs/src/share/get-file-stats.ts
@@ -1,0 +1,23 @@
+import { type Stats } from 'node:fs';
+import { stat as statAsync } from 'node:fs/promises';
+
+/**
+ * Get file stats with error handling
+ */
+export async function getFileStats(filePath: string): Promise<
+  | {
+      success: true;
+      stats: Stats;
+    }
+  | {
+      success: false;
+      error: unknown;
+    }
+> {
+  try {
+    const stats = await statAsync(filePath);
+    return { success: true, stats };
+  } catch (error) {
+    return { success: false, error };
+  }
+}
