+++ packages/serve-static-plugin-nodejs/src/version.ts
