+++ packages/serve-static-plugin-nodejs/tsconfig.json
