+++ packages/serve-static-plugin-nodejs/tsdown.config.js
@@ -6,5 +6,5 @@ export default defineConfig({
   dts: true,
   clean: false,
   sourcemap: true,
-  external: ['@korix/kori', 'node:fs', 'node:fs/promises', 'node:path', 'node:crypto', 'mime-types'],
+  external: ['@korix/kori', 'node:fs', 'node:fs/promises', 'node:path', 'etag', 'mime-types'],
 });
