--- packages/static-file-plugin-nodejs/src/index.ts
@@ -1,2 +1,0 @@
-export { type StaticFileOptions } from './static-file-options.js';
-export { staticFilePlugin } from './static-file-plugin.js';
