--- packages/static-file-plugin-nodejs/src/static-file-plugin.ts
@@ -1,410 +1,0 @@
-import {
-  defineKoriPlugin,
-  type KoriRequest,
-  type KoriResponse,
-  type KoriLogger,
-  type KoriEnvironment,
-  type KoriPlugin,
-  HttpStatus,
-  HttpResponseHeader,
-} from '@korix/kori';
-import mime from 'mime-types';
-
-import {
-  resolveSafePath,
-  getFileInfo,
-  resolveIndexFile,
-  isDotfileAllowed,
-  generateETag,
-  formatLastModified,
-  createFileStream,
-  createPartialFileStream,
-  parseRangeHeader,
-  generateContentRangeHeader,
-  generateBoundary,
-  createMultipartStream,
-  type ExistingFileInfo,
-} from './file-utils.js';
-import { type StaticFileOptions, RangeConstants, type RangeResult } from './static-file-options.js';
-import { PLUGIN_VERSION } from './version.js';
-
-const PLUGIN_NAME = 'static-file-plugin-nodejs';
-
-/**
- * MIME type detection using mime-types library for comprehensive coverage
- */
-function detectMimeType(filePath: string): string {
-  const mimeType = mime.lookup(filePath);
-  return mimeType || 'application/octet-stream';
-}
-
-const defaultOptions: Required<Omit<StaticFileOptions, 'serveFrom'>> = {
-  mountAt: '/static',
-  index: ['index.html'],
-  dotfiles: 'deny',
-  maxAge: 0,
-  etag: true,
-  lastModified: true,
-  immutable: false,
-  ranges: true,
-  maxRanges: 1,
-};
-
-function validateOptions(options: StaticFileOptions, log: KoriLogger): void {
-  if (!options.serveFrom) {
-    const errorMessage = 'Static file plugin requires a serveFrom directory';
-    log.error(errorMessage);
-    throw new Error(errorMessage);
-  }
-
-  if (options.maxAge !== undefined && options.maxAge < 0) {
-    const errorMessage = 'maxAge must be a non-negative number';
-    log.error(errorMessage, { maxAge: options.maxAge });
-    throw new Error(errorMessage);
-  }
-
-  if (options.mountAt && !options.mountAt.startsWith('/')) {
-    const errorMessage = 'mountAt must start with "/"';
-    log.error(errorMessage, { mountAt: options.mountAt });
-    throw new Error(errorMessage);
-  }
-}
-
-function removeMountPrefix(pathname: string, mountAt: string): string {
-  if (pathname === mountAt) {
-    return '/';
-  }
-
-  if (!pathname.startsWith(mountAt)) {
-    throw new Error(
-      `[INTERNAL] Pathname "${pathname}" does not start with mount prefix "${mountAt}". This indicates a routing system bug.`,
-    );
-  }
-
-  return pathname.slice(mountAt.length);
-}
-
-function setCacheHeaders(
-  res: KoriResponse,
-  fileInfo: ExistingFileInfo,
-  options: Required<Pick<StaticFileOptions, 'maxAge' | 'etag' | 'lastModified' | 'immutable'>>,
-): void {
-  if (options.maxAge > 0) {
-    let cacheControl = `public, max-age=${options.maxAge}`;
-    if (options.immutable) {
-      cacheControl += ', immutable';
-    }
-    res.setHeader(HttpResponseHeader.CACHE_CONTROL, cacheControl);
-  } else {
-    res.setHeader(HttpResponseHeader.CACHE_CONTROL, 'no-cache');
-  }
-  if (options.etag) {
-    const etag = generateETag(fileInfo.stats);
-    res.setHeader(HttpResponseHeader.ETAG, etag);
-  }
-  if (options.lastModified) {
-    const lastModified = formatLastModified(fileInfo.stats.mtime);
-    res.setHeader(HttpResponseHeader.LAST_MODIFIED, lastModified);
-  }
-}
-
-function checkConditionalRequest(req: KoriRequest, fileInfo: ExistingFileInfo): boolean {
-  // ETag-based conditional request (priority)
-  const ifNoneMatch = req.header('if-none-match');
-  if (ifNoneMatch) {
-    const etag = generateETag(fileInfo.stats);
-    return ifNoneMatch === etag;
-  }
-
-  // Last-Modified-based conditional request (fallback)
-  const ifModifiedSince = req.header('if-modified-since');
-  if (ifModifiedSince) {
-    try {
-      // HTTP dates have second precision, so truncate to seconds for comparison
-      const fileModifiedTime = Math.floor(fileInfo.stats.mtime.getTime() / 1000);
-      const clientModifiedTime = Math.floor(new Date(ifModifiedSince).getTime() / 1000);
-
-      // If file hasn't been modified since client's timestamp, return 304
-      return fileModifiedTime <= clientModifiedTime;
-    } catch {
-      // Invalid date format in If-Modified-Since header, ignore
-      return false;
-    }
-  }
-
-  return false;
-}
-
-function serveFile(
-  req: KoriRequest,
-  res: KoriResponse,
-  fileInfo: ExistingFileInfo,
-  options: Required<Pick<StaticFileOptions, 'maxAge' | 'etag' | 'lastModified' | 'immutable' | 'ranges' | 'maxRanges'>>,
-  log: KoriLogger,
-): KoriResponse {
-  const mimeType = detectMimeType(fileInfo.path);
-
-  // Check for conditional request first
-  if (checkConditionalRequest(req, fileInfo)) {
-    log.debug('Serving 304 Not Modified', { path: fileInfo.path });
-    return res.status(HttpStatus.NOT_MODIFIED).empty();
-  }
-
-  // Check if this is a range request and ranges are enabled
-  const rangeHeader = req.header('range');
-  if (options.ranges && rangeHeader) {
-    const rangeResult = parseRangeHeader(rangeHeader, fileInfo.stats.size);
-    if (rangeResult.isSatisfiable && rangeResult.ranges.length > 0) {
-      return serveRangeRequest(req, res, fileInfo, options, log, rangeResult);
-    } else {
-      // Handle unsatisfiable ranges
-      log.debug('Range not satisfiable', {
-        path: fileInfo.path,
-        rangeHeader,
-        fileSize: fileInfo.stats.size,
-      });
-
-      res.setHeader(HttpResponseHeader.CONTENT_RANGE, `${RangeConstants.BYTES} */${fileInfo.stats.size}`);
-      res.setHeader(HttpResponseHeader.CONTENT_TYPE, mimeType);
-
-      return res.status(HttpStatus.RANGE_NOT_SATISFIABLE).json({
-        error: {
-          type: 'RANGE_NOT_SATISFIABLE',
-          message: 'Requested range not satisfiable',
-        },
-      });
-    }
-  }
-
-  res.setHeader(HttpResponseHeader.CONTENT_TYPE, mimeType);
-  res.setHeader(HttpResponseHeader.CONTENT_LENGTH, fileInfo.stats.size.toString());
-  res.setHeader(HttpResponseHeader.ACCEPT_RANGES, options.ranges ? RangeConstants.BYTES : RangeConstants.NONE);
-  setCacheHeaders(res, fileInfo, options);
-
-  log.debug('Serving complete file', {
-    path: fileInfo.path,
-    size: fileInfo.stats.size,
-    mimeType,
-  });
-
-  const fileStream = createFileStream(fileInfo.path);
-  return res.status(HttpStatus.OK).stream(fileStream);
-}
-
-function serveRangeRequest(
-  req: KoriRequest,
-  res: KoriResponse,
-  fileInfo: ExistingFileInfo,
-  options: Required<Pick<StaticFileOptions, 'maxAge' | 'etag' | 'lastModified' | 'immutable' | 'ranges' | 'maxRanges'>>,
-  log: KoriLogger,
-  rangeResult: RangeResult,
-): KoriResponse {
-  const mimeType = detectMimeType(fileInfo.path);
-  const fileSize = fileInfo.stats.size;
-
-  // Check if requesting too many ranges
-  if (rangeResult.ranges.length > options.maxRanges) {
-    log.warn('Too many ranges requested', {
-      path: fileInfo.path,
-      requestedRanges: rangeResult.ranges.length,
-      maxRanges: options.maxRanges,
-    });
-
-    return res.status(HttpStatus.RANGE_NOT_SATISFIABLE).json({
-      error: {
-        type: 'TOO_MANY_RANGES',
-        message: `Too many ranges requested. Maximum allowed: ${options.maxRanges}`,
-      },
-    });
-  }
-
-  // Handle single vs multiple range requests
-  if (rangeResult.ranges.length === 1) {
-    // Single range request - ranges[0] is guaranteed to exist since length === 1
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const range = rangeResult.ranges[0]!;
-    return serveSingleRange(req, res, fileInfo, options, log, range, mimeType, fileSize);
-  } else {
-    // Multiple range request - use multipart response
-    return serveMultipartRange(req, res, fileInfo, options, log, rangeResult.ranges, mimeType, fileSize);
-  }
-}
-
-function serveSingleRange(
-  req: KoriRequest,
-  res: KoriResponse,
-  fileInfo: ExistingFileInfo,
-  options: Required<Pick<StaticFileOptions, 'maxAge' | 'etag' | 'lastModified' | 'immutable' | 'ranges' | 'maxRanges'>>,
-  log: KoriLogger,
-  range: { start: number; end: number },
-  mimeType: string,
-  fileSize: number,
-): KoriResponse {
-  const contentLength = range.end - range.start + 1;
-
-  log.debug('Serving single range request', {
-    path: fileInfo.path,
-    range: `${range.start}-${range.end}`,
-    contentLength,
-    totalSize: fileSize,
-    mimeType,
-  });
-
-  // Set response headers for partial content
-  res.setHeader(HttpResponseHeader.CONTENT_TYPE, mimeType);
-  res.setHeader(HttpResponseHeader.CONTENT_LENGTH, contentLength.toString());
-  res.setHeader(HttpResponseHeader.CONTENT_RANGE, generateContentRangeHeader(range.start, range.end, fileSize));
-  res.setHeader(HttpResponseHeader.ACCEPT_RANGES, RangeConstants.BYTES);
-
-  // Set cache headers
-  setCacheHeaders(res, fileInfo, options);
-
-  // Create partial file stream
-  const partialStream = createPartialFileStream(fileInfo.path, range.start, range.end);
-
-  return res.status(HttpStatus.PARTIAL_CONTENT).stream(partialStream);
-}
-
-function serveMultipartRange(
-  req: KoriRequest,
-  res: KoriResponse,
-  fileInfo: ExistingFileInfo,
-  options: Required<Pick<StaticFileOptions, 'maxAge' | 'etag' | 'lastModified' | 'immutable' | 'ranges' | 'maxRanges'>>,
-  log: KoriLogger,
-  ranges: { start: number; end: number }[],
-  mimeType: string,
-  fileSize: number,
-): KoriResponse {
-  const boundary = generateBoundary();
-
-  log.debug('Serving multipart range request', {
-    path: fileInfo.path,
-    rangeCount: ranges.length,
-    ranges: ranges.map((r: { start: number; end: number }) => `${r.start}-${r.end}`),
-    totalSize: fileSize,
-    mimeType,
-    boundary,
-  });
-
-  // Set response headers for multipart content
-  // Note: Content-Length is omitted for multipart responses as per HTTP spec
-  // since calculating exact length is complex and not required
-  res.setHeader(HttpResponseHeader.CONTENT_TYPE, `multipart/byteranges; boundary=${boundary}`);
-  res.setHeader(HttpResponseHeader.ACCEPT_RANGES, RangeConstants.BYTES);
-
-  // Set cache headers
-  setCacheHeaders(res, fileInfo, options);
-
-  // Create multipart stream
-  const multipartStream = createMultipartStream(fileInfo.path, ranges, fileSize, mimeType, boundary);
-
-  return res.status(HttpStatus.PARTIAL_CONTENT).stream(multipartStream);
-}
-
-async function handleStaticFileRequest(
-  req: KoriRequest,
-  res: KoriResponse,
-  requestPath: string,
-  options: Required<StaticFileOptions>,
-  log: KoriLogger,
-): Promise<KoriResponse> {
-  const resolvedPath = resolveSafePath(requestPath, options.serveFrom);
-
-  if (!resolvedPath.isValid) {
-    log.warn('Invalid file path detected', { requestPath });
-    return res.notFound({
-      message: 'File not found',
-    });
-  }
-
-  if (!isDotfileAllowed(resolvedPath.isDotfile, options.dotfiles)) {
-    log.debug('Dotfile access denied', { path: resolvedPath.safePath });
-    return res.notFound({
-      message: 'File not found',
-    });
-  }
-
-  let fileInfo = await getFileInfo(resolvedPath.safePath);
-
-  if (!fileInfo.exists) {
-    log.debug('File not found', { path: resolvedPath.safePath });
-    return res.notFound({
-      message: 'File not found',
-    });
-  }
-
-  if (fileInfo.stats.isDirectory()) {
-    if (options.index === false) {
-      log.debug('Directory listing disabled', { path: resolvedPath.safePath });
-      return res.forbidden({
-        message: 'Directory listing is disabled',
-      });
-    }
-
-    const indexFile = await resolveIndexFile(resolvedPath.safePath, options.index, log);
-    if (!indexFile) {
-      log.debug('No index file found in directory', { path: resolvedPath.safePath });
-      return res.notFound({
-        message: 'No index file found',
-      });
-    }
-    fileInfo = indexFile;
-  }
-
-  if (!fileInfo.stats.isFile()) {
-    log.debug('Path is not a regular file', { path: fileInfo.path });
-    return res.notFound({
-      message: 'File not found',
-    });
-  }
-
-  return serveFile(req, res, fileInfo, options, log);
-}
-
-/**
- * Static file serving plugin for Node.js
- *
- * Features:
- * - Secure file serving with path traversal protection
- * - Streaming file delivery for efficient memory usage
- * - MIME type detection for proper Content-Type headers
- * - Caching support with ETag and Last-Modified headers
- * - Configurable cache max-age
- * - Index file resolution for directory requests
- * - Dotfiles handling
- */
-export function staticFilePlugin<Env extends KoriEnvironment, Req extends KoriRequest, Res extends KoriResponse>(
-  userOptions: StaticFileOptions,
-): KoriPlugin<Env, Req, Res> {
-  const options = {
-    ...defaultOptions,
-    ...userOptions,
-  };
-
-  return defineKoriPlugin({
-    name: PLUGIN_NAME,
-    version: PLUGIN_VERSION,
-    apply: (kori) => {
-      const log = kori.log().child(PLUGIN_NAME);
-
-      validateOptions(options, log);
-
-      log.info('Static file plugin initialized', {
-        serveFrom: options.serveFrom,
-        mountAt: options.mountAt,
-        index: options.index,
-        dotfiles: options.dotfiles,
-        maxAge: options.maxAge,
-        etag: options.etag,
-        lastModified: options.lastModified,
-      });
-
-      return kori.get(`${options.mountAt}/*`, async ({ req, res }) => {
-        const pathname = req.url().pathname;
-        const requestPath = removeMountPrefix(pathname, options.mountAt);
-
-        return await handleStaticFileRequest(req, res, requestPath, options, log);
-      });
-    },
-  });
-}
