+++ tsconfig.json
@@ -14,7 +14,7 @@
     { "path": "./packages/script" },
     { "path": "./packages/security-headers-plugin" },
     { "path": "./packages/send-file-plugin-nodejs" },
-    { "path": "./packages/static-file-plugin-nodejs" },
+    { "path": "./packages/serve-static-plugin-nodejs" },
     { "path": "./packages/zod-openapi-plugin" },
     { "path": "./packages/zod-schema" },
     { "path": "./packages/zod-validator" }
